Rename progexec_thread.py
* Rename progexec_thread.py to machif_progexec.py
* Fix bug with double key on g2core error dictionary
