+++ modules/g2core_machif.py
@@ -193,7 +193,7 @@ G2CORE_STAT_CODE_2_STR_DICT = {
     228:"SOFT_LIMIT_EXCEEDED_AMAX",
     229:"SOFT_LIMIT_EXCEEDED_BMIN",
 
-    220:"SOFT_LIMIT_EXCEEDED_BMAX",
+    230:"SOFT_LIMIT_EXCEEDED_BMAX",
     231:"SOFT_LIMIT_EXCEEDED_CMIN",
     232:"SOFT_LIMIT_EXCEEDED_CMAX",
     233:"SOFT_LIMIT_EXCEEDED_ARC",
